net/mlx5/hws: supporting default miss table in HWS
A default miss table is a way to define what happens to traffic that does
not match any rule in a specific table. In hws, this is done by connecting
the source table to the target table using the RTC.
This ensures that traffic that does not match any rule in the source table
is forwarded to the target table.
Issue: 3441251

Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>