eal: add flag to indicate non-EAL malloc heaps
Rather than relying implicitly on the socket_id value to identify
external heaps vs internal ones that can be expanded on demand by adding
more hugepages, we add an "is_external" flag to the heap structure. As
we do so, we change the heap initialization to use designated
initializers to guarantee as we add new fields that they are properly
zeroed on init. As it stands, many fields are not explicitly set on
first init of the internal malloc heaps.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>