config/arm: update aarch32 build with gcc13
The aarch32 with gcc13 fails with:

Compiler for C supports arguments -march=armv8-a: NO

../config/arm/meson.build:714:12: ERROR: Problem encountered: No
suitable armv8 march version found.

This is because we test -march=armv8-a without the -mpfu option, which
is required for aarch32 build with gcc13.

On top of that, the most recent recommendation from the compiler team to
build with -march=armv8-a+simd -mfpu=auto.

We can address this by first testing compiler options and only test the
march after that (coupled with supported options). This doesn't fully
work though, as just -march=armv8 in not enough to be accepted with
-mfpu=auto. To address this other issue, add the force_march parameter
to part number config which will force the configured march to be used,
even if the compiler doesn't support the particular combination of march
and other compiler options, such as mfpu. This then allows the check to
proceed to the -march=armv8-a+simd -mfpu=auto, which is valid.

Signed-off-by: Juraj Linkeš <juraj.linkes@pantheon.tech>
Signed-off-by: 0-day Robot <robot@bytheb.org>