net/nfp: update incorrect MAC stats offset
The pointer to the beginning of the MAC stats counters for port 1 are
incorrectly set as the pointer to the beginning of the port 0 MAC stats
counters, plus the size of the MAC stats counters multiplied by the port
number.

This patch corrects this by setting the multiplier as the eth_table
index of the port.

Fixes: f26e82397f6d ("net/nfp: implement xstats")
Cc: stable@dpdk.org

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Peng Zhang <peng.zhang@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>