doc: remove confusing command to send patch
In the contributor guide, it was said that no need to Cc maintainers
for new additions, probably for new directories not having a maintainer.
There is no harm, and it is a good habit, to always Cc maintainers.

Remove this case as it can mislead to not Cc maintainers when needed.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>