eal: disable standard c atomics support check for MSVC
MSVC will define __STDC_NO_ATOMICS__ until versions of the compiler that
support atomics is released. Currently we use the Preview version of the
compiler so skip the test of __STDC_NO_ATOMICS__ avoiding failure.

This is a temporary change until the required compiler is released
publicly but allows us to establish the MSVC CI pipeline.

Signed-off-by: Tyler Retzlaff <roretzla@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>