ice: Increase L1 check interval to account for lengthier link time
For edge cases where the transceiver is physically inserted first and immediately afterwards the DPDK PF is started the LSC event may occur outside the current setting for the maximum check interval window. This change lengthens the check interval to account for this along with other reported cases where the link event be longer than 1 second.

Cc: stable@dpdk.org

Reported-by: Timothy Miskell
Suggested-by: Timothy Miskell
Signed-off-by: Timothy Miskell
Reviewed-by: Larry Wang
Tested-by: Jonathan Tsai
Signed-off-by: 0-day Robot <robot@bytheb.org>