net/ice: fix L1 check interval to account for longer link times
For edge cases where the transceiver is physically inserted first and
immediately afterwards the DPDK PF is started the LSC event may occur
outside the current setting for the maximum check interval window. This
change lengthens the check interval to account for this along with other
reported cases where the link event may be longer than 1 second.

Signed-off-by: Timothy Miskell <timothy.miskell@intel.com>
Tested-by: Jonathan Tsai <jonathan1.tsai@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>