usertools/rss: add driver abstractions
The default RETA size is not the same for all drivers. In some drivers
(mlx5), the RETA size may also be dependent on the number of RX queues.

Introduce a new DriverInfo abstraction for known keys. Define a simple
API to expose the RSS key and RETA size (based on the number of RX
queues).

Use that abstraction for all three known keys.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>