node: remove some of the experimental tags
The node library has been around long enough that experimental
tag should be removed.  The IPv6 support was added in 23.03
but that is still enough time that it can be marked stable.

The logtype variable should have been marked internal
since it is not intended to be used outside of library.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>